Refactored code

+++ BlackJack.DAL/Interfaces/ICardRepository.cs
@@ -6,7 +6,7 @@ namespace BlackJack.DataAccess.Interfaces
 {
 	public interface ICardRepository
 	{
-		Task<List<Card>> GetCardsById(List<int> cardsId);
+		Task<List<Card>> GetCardsById(List<long> cardsId);
 		Task PopulateCards(List<Card> cards);
 		Task<List<Card>> GetAll();
 		Task DeleteAll();
