+++ BlackJack.DAL/Interfaces/IHandRepository.cs
@@ -5,9 +5,9 @@ namespace BlackJack.DataAccess.Interfaces
 {
 	public interface IHandRepository
     {
-        Task GiveCardToPlayerInGame(int playerId, int cardId, int gameId);
-        Task<List<int>> GetCardsIdByPlayerId(int playerId, int gameId);
-        Task RemoveAllCardsInHand(int gameId);
-		Task<List<int>> GetCardsIdByGameId(int gameId);
+        Task GiveCardToPlayerInGame(long playerId, long cardId, long gameId);
+        Task<List<long>> GetCardsIdByPlayerId(long playerId, long gameId);
+        Task RemoveAllCardsInHand(long gameId);
+		Task<List<long>> GetCardsIdByGameId(long gameId);
 	}
 }
