+++ BlackJack.DAL/Interfaces/IPlayerInGameRepository.cs
@@ -6,17 +6,17 @@ namespace BlackJack.DataAccess.Interfaces
 {
 	public interface IPlayerInGameRepository
     {
-        Task AddPlayerToGame(int playerId, int gameId, bool isHuman);
-        Task RemoveAllPlayersFromGame(int gameId);
-        Task<List<int>> GetAllPlayersIdByGameId(int gameId);
-        Task<List<int>> GetBotsIdByGameId(int gameId);
-        Task<int> GetBetByPlayerId(int playerId, int gameId);
-        Task PlaceBet(int playerId, int bet, int gameId);
-		Task PlaceBet(List<int> playerId, int gameId);
-		Task AnnulBet(int playerId, int gameId);
-		Task AnnulBet(List<int> playersId, int gameId);
-		Task<int> IsInGame(int playerId, int gameId);
-		Task<List<PlayerInGame>> GetPlayersInGame(List<int> playersId, int gameId);
-		Task<int> GetHumanIdByGameId(int gameId);
+        Task AddPlayerToGame(long playerId, long gameId, bool isHuman);
+        Task RemoveAllPlayersFromGame(long gameId);
+        Task<List<long>> GetAllPlayersIdByGameId(long gameId);
+        Task<List<long>> GetBotsIdByGameId(long gameId);
+        Task<int> GetBetByPlayerId(long playerId, long gameId);
+        Task PlaceBet(long playerId, int bet, long gameId);
+		Task PlaceBet(List<long> playerId, long gameId);
+		Task AnnulBet(long playerId, long gameId);
+		Task AnnulBet(List<long> playersId, long gameId);
+		Task<long> IsInGame(long playerId, long gameId);
+		Task<List<PlayerInGame>> GetPlayersInGame(List<long> playersId, long gameId);
+		Task<long> GetHumanIdByGameId(long gameId);
 	}
 }
