+++ BlackJack.Entity/BaseEntity.cs
@@ -6,7 +6,7 @@ namespace BlackJack.Entities
 	public class BaseEntity
 	{
 		[Key]
-		public int Id { get; set; }
+		public long Id { get; set; }
 		public DateTime CreationDate { get; set; }
 		
 		public BaseEntity()
@@ -14,4 +14,4 @@ namespace BlackJack.Entities
 			CreationDate = DateTime.Now;
 		}
 	}
-}
+}
